hypecart_agents/requirements.txt
google-adk
Flask
mcp
